Background
the background for level_scene.tscn pretty mid honestly, I've been super sick lately and i don't know when I'll be able to work on stuff. Super low poly stuff roughly 35-40 thousand tris total, the world sphere that has the sky does look dark for no reason idk how to solve that. If this is too bad, hopefully I can get to work in a pre-rendered alternative like I had it at first but with Skin's color scheme and all that, and to fit the new overall positioning, if that makes any sense.